modularize lneto/gvisor stacks
